Don't run migration code that modifies the users table at all if a User constant exists
Based off ideas from this comment: https://github.com/spree/spree/pull/1512#issuecomment-6136557

+++ core/db/migrate/20101103212716_drop_anonymous_field_for_user.rb
@@ -1,6 +1,8 @@
 class DropAnonymousFieldForUser < ActiveRecord::Migration
   def up
-    remove_column :users, :anonymous
+    unless defined?(User)
+      remove_column :users, :anonymous
+    end
   end
 
   def down
